pyproject.toml
[project]
name = "cava_llm_manager"
version = "0.1.0"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "pydantic>=2",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
